sweet and eazy back and front configuration in the same repo
