+++ back/main.ts
@@ -1,0 +1,3 @@
+import { log } from '../common/mylib';
+
+log("Imported a common lib. And server is running");
