+++ common/mylib.ts
@@ -1,0 +1,3 @@
+export function log(value: string) {
+    console.log(value);
+}
