+++ front/components/box.test.tsx
@@ -1,0 +1,5 @@
+import { render, screen } from '@testing-library/react';
+
+test('renders learn react link', () => {
+  expect(1).toBe(1);
+});
