+++ front/vite.config.js
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [react()],
+})
