+++ package.json
@@ -1,48 +1,32 @@
 {
   "name": "capt",
-  "version": "0.1.0",
   "private": true,
+  "version": "0.0.0",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "front": "vite serve front --config front/vite.config.js",
+    "back": "node back/dist/bundle.back.js",
+    "build:front": "vite build front --config front/vite.config.js",
+    "build:back": "rollup --config back/rollup.config.ts",
+    "test:front": "jest --config front/jest.config.ts"
   },
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.1.1",
-    "@testing-library/user-event": "^13.5.0",
-    "konva": "^8.3.6",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "react-konva": "^18.0.0-0",
-    "react-markdown": "^8.0.3",
-    "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4"
+    "@rollup/plugin-typescript": "^8.3.2",
+    "konva": "^8.3.8",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-konva": "^18.1.0",
+    "react-markdown": "^8.0.3"
   },
   "devDependencies": {
-    "@types/jest": "^27.4.1",
-    "@types/node": "^17.0.30",
-    "@types/react": "^18.0.8",
-    "@types/react-dom": "^18.0.3",
-    "typescript": "^4.6.4"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "@testing-library/react": "^13.2.0",
+    "@types/jest": "^27.5.1",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@vitejs/plugin-react": "^1.3.0",
+    "jest": "^28.1.0",
+    "ts-jest": "^28.0.2",
+    "ts-node": "^10.8.0",
+    "typescript": "^4.6.3",
+    "vite": "^2.9.9"
   }
 }
