--- src/types/arrow.ts
@@ -1,7 +1,0 @@
-import { Tip } from "./tip";
-
-export type Arrow = {
-  id: number,
-  start: Tip,
-  end: Tip
-}
