--- src/types/projection.ts
@@ -1,10 +1,0 @@
-export type Projection = {
-  inside: boolean,
-  dist: number,
-  id: number, 
-  percentage: number,
-  px: number,
-  py: number,
-  side: string
-}
-
